videosource: Try reseeking when av_read_frame() returns EOF
In rare cases, an EOF error may be returned prematurely in the decode loop.
This can happen when processing an index generated with the option use_mfra_for,
if the source file contains invalid fragment offsets between valid ones in the tfra box.
Forcing a seek can bypass those invalid fragments and allow decoding to proceed normally.
A frame check is added to prevent repeating this operation if it was already attempted on the same frame.